refactor(bootstrap): one CRUD map on the live path, preflight becomes a facade (#124)
Retires the duplicate map. Two copies existed: bootstrap/preflight/ carried CRUD
depth with no production consumer, while bootstrap/resource-action-map.ts was
create-only and wired into the live synth-coverage gate. Disjoint test suites and
no shared consumer means they drift by construction, and adding a resource type
to only one of them is silent.

Merged programmatically, not by hand, with the invariant asserted mechanically:
every action from the create-only map survives in the merged entry's `create`
phase (verified 0 lost across 52 types). Result is 64 types / 430 actions, up
from 52 create-only entries — the CRUD map contributed 48 create-phase actions
the live map lacked on shared types, plus AWS::IAM::ManagedPolicy, while main's
6 extra types (CloudFront, Custom::*, CDK::Metadata) are preserved.

- RESOURCE_ACTION_MAP is now Record<string, ResourceActions> with
  create/read/update/delete. findMissingBootstrapActions defaults to ['create'],
  preserving the pre-CRUD contract for existing callers; pass phases to widen.
- bootstrap/preflight/resource-action-map.ts holds NO data — it re-exports the
  single map and keeps the query helpers (getActionsForResource,
  getAllMappedActions) that #125/#126 will read it through. 428 lines -> 62.
- Deleted the two vacuous 'Synth coverage' tests here: both bailed silently
  (`catch { return }`, `types.length === 0`) and the ECS one burned ~86s
  asserting nothing. synth-coverage.test.ts now covers both configs in-process
  and fails loudly (previous commit).

KNOWN_GAP_SERVICES / KNOWN_GAP_ACTIONS removed entirely. Verified every one of
the 11 excluded actions is now covered — the sqs/s3 service-wide exclusions and
all 7 lambda actions were stale, hiding nothing. With the 4 real gaps granted in
the previous commit, the coverage assertion runs over ALL 430 actions in ALL four
phases with zero exemptions, which is what krokoko asked for ("the test passes
only by excluding the cases it most needs to catch").

Mutation-tested: revoking sqs:AddPermission fails with
"1 actions not covered by bootstrap policies: sqs:AddPermission".

Added structural pins so the depth cannot erode: every entry must declare all
four phases as arrays, and >=45 entries must carry real update/delete actions.

178 suites / 3533 tests pass; //cdk:eslint clean, no mutations.
